Add a Japanese Playground Docs translation. (#2198)
## Motivation for the change, related issues
related issues #1798

Hello, I'm making a PR to add a Japanese translation of the Playground
Docs.

There are a lot of WordPress users in Japan. I hope that adding the
translation of the documentation will help them with their development.

## Implementation details

I added a Japanese translation according to this pull request.
#1766

I was able to check the Japanese documentation locally.


![playground-ja](https://github.com/user-attachments/assets/991830fb-fb48-41f9-b2d4-3f48f763f5bf)