[Plugin Proxy] Allow all WordPress, Automattic and Woocommerce repos (#128)
## Motivation for the change, related issues

Currently, the proxy is only open to very specific repos that was done
to avoid abuse when the public playground instance was used. Now that we
have a private one that we are in control of, we can open this up to all
Automattic, WordPress and Woocommerce repos. Those are trusted
Automattic repos that we control and can be assured of their security.

## Implementation details

There is a list of orgs that are allowed, and only the repos that match
these orgs are okay within the proxy. Anything else is rejected.

I've also tried to limit what's exposed via any exceptions that are
thrown for security reasons.

## Testing Instructions (or ideally a Blueprint)

- Run a php server within the public folder
- Attempt to test against an Automattic plugin that would previously not
be allowed, like Remote Data Blocks

---------

Co-authored-by: Adam Zieliński <adam@adamziel.com>