Remove null steps that are added (#135)
## Motivation for the change, related issues

The
[resolveBlueprintFromURL](https://github.com/Automattic/wp-calypso/blob/99caa52a8105ebbe0ad78e911665815ca522df2e/client/landing/stepper/declarative-flow/internals/steps-repository/playground/lib/resolve-blueprint-from-url.ts#L64-L89)
function returns a steps: `[null, null, null]` array. This happens when
`importWxrQueryArg` is `null` or `query.get( 'import-site' )` is `null`
or `query.get( 'theme' )` is `null`.

If these values are `null` the step is added as a `null` value. It
should not be added at all.


https://linear.app/a8c/issue/DOTOBRD-124/wrong-blueprints-steps-from-resolveblueprintfromurl-url-url

Props to @zaerl for the investigation

## Implementation details

- Filter falsy value out (easier to just look at both commits)