Avoid undefined __dirname error when running unbuilt Playground CLI in Node.js (#2233)
## Motivation for the change, related issues

We currently use `__dirname` to resolve ICU data, but that does not work
for running unbuilt TS modules in Node.js (for example, via `npx nx
dev-node playground-cli`). In that case, we get an error that
`__dirname` not defined because the global does not exist for ES modules
in Node.

This PR fixes this so the path can be resolved by both.

cc @mho22 

## Implementation details

This PR updates the unbuilt module loader to replace uses of `__dirname`
and `__filename`. Some notes:
- `__dirname` is replaced by `import.meta.dirname` which is supported
for ES modules at least as early as Node.js 20.x.
- `__filename` is replaced by `import.meta.filename` which is supported
for ES modules at least as early as Node.js 20.x.
- This is a text-only replacement, which is fast but not as ideal as an
AST transform.
- The text-only replacement avoids replacing declarations of `__dirname`
and `__filename` because that would cause a syntax error.

This PR also removes [the conditional declarations of `__dirname` and
`__filename`](https://github.com/WordPress/wordpress-playground/blob/8ccf4e47f9d84401c8f237ae9b039060a45a0f8a/packages/php-wasm/compile/php/esm-node-boilerplate.js#L8-L21)
from `esm-node-boilerplate.js` because they are no longer needed.

## Testing Instructions (or ideally a Blueprint)

We can test this with CI.

There is a new `test-running-unbuilt-cli.sh` script for running an
unbuilt version of Playground CLI, and it is integrated in the CI
workflow as a standalone job.