Adding Brazilian Portuguese Intro documentation (#2307)
## Motivation for the change, related issues

This pull request adds the necessary structure to translate the
WordPress Playground documentation into Brazilian Portuguese, as the
first step towards resolving issue #2306.

## Implementation details

The `pt-BR` folder follows the structure from other languages, adding
the JSON files for localization of components and the `intro.md` file as
the first translation.

## Testing 

To run it locally:
1. Download the repository
2. Go to the terminal and run `npm install`
3. Run the command at the terminal: `npm run dev:docs -- --locale pt-BR`