Make fs-ext dependency optional (#2322)
## Motivation for the change, related issues

Some folks are encountering issue with building the fs-ext module
containing a native Node.js add-on.

Fixes #2308

## Implementation details

This PR makes the fs-ext dependency optional. This means that the
dependency will be proactively installed by `npm`, but if an optional
dependency fails to install, it does not fail the entire npm install.

## Testing Instructions (or ideally a Blueprint)

- Run `npm ci`
- Confirm that `fs-ext` is installed under node_modules
- Edit packages/php-wasm/node/src/lib/file-lock-manager-for-node.ts and
add a console.log() statement to `flockSyncNoOp()`
- Make local dir test dir
- Run `npx nx unbuilt-jspi playground-cli server
--mountBeforeInstall=<local-test-dir-path>:/wordpress`
- Confirm it starts with no errors and no `flockSyncNoOp` messages.
- Delete `node_modules/fs-ext`
- Run `npx nx unbuilt-jspi playground-cli server
--mountBeforeInstall=<local-test-dir-path>:/wordpress`
- Confirm it starts successfully and prints `flockSyncNoP` console
messages.