[Build] Preserve optionalDependencies in built package.json (#2410)
## Motivation for the change, related issues

The `fs-ext` package is listed under `"optionalDependencies"` in the
top-level package.json. However, in the built packages, it becomes a
regular `dependency`. This PR makes sure that all the
`optionalDependencies` remain `optionalDependencies` in the final built
package.

## Testing Instructions (or ideally a Blueprint)

Run `npm run build` and confirm that the built `@wp-playground/cli` and
`@php-wasm/node` list `fs-ext` under `optionalDependencies` and not
under `dependencies`.