Playground CLI: Wire up execution modes for Blueprints v2  (#2519)
## Motivation for the change, related issues

Applying a Blueprint to an existing site is not possible with the
Blueprints v2 runner as it always runs in the `create-new-site` mode,
not in the `apply-to-existing-site` mode. This PR exposes an explicit
`--mode` option

Related to #2508

## Implementation details

This PR exposes a `mode` CLI option for use with Blueprints v2.

For migration convenience, this PR also rewrites a few v1 options (e.g.
`--skipWordpressSetup`) into v2 options (e.g.
`--mode=apply-to-existing-site`). There are appropriate safeguard in
place to prevent options conflicts.

In the categories of bonus, controversial, and up for discussion, this
PR also expands the auto-mount option to allow passing an explicit path
to auto-mount. If no path is passed the option defaults the current
working directory. This is a nice feature IMO, and it also makes testing
the CLI feature using `npx nx unbuilt-jspi playground-cli` (since the
current directory for that command is typically the project root).

## Testing Instructions (or ideally a Blueprint)

- CI

---------

Co-authored-by: Adam Zieliński <adam@adamziel.com>