Updating runCLI demo at wp-playground/cli README file (#2553)
## Motivation for the change, related issues
The `runCLI` demo is showing an outdated demo; the flow using the runCLI
has changed, and the example displayed at the latest playground package
doesn't run. This pull request updates the demo with a simple example to
spin a playground CLI server.

This pull request is part of the issue #2552