[Docs] Adding how to contribute to the WordPress Playground (#2597)
## Motivation for the change, related issues

This pull request updates the Contributor Day page in the documentation,
including an intro about what Contributor Day is and how contributors
can help with the project. With this change, the page is organized into
two blocks: "How you can contribute to Playground" and "How Playground
can support contributors during the Contributor Day".

The goal is to guide contributors to areas where they can collaborate
with the project and to provide guidance on how to continue contributing
to the project after it has been completed.

A second page will be created for Playground table leads