[Website] Report Blueprint v2 progress (#2674)
## Motivation for the change, related issues

Captures the Blueprint v2 execution progress events and communicates the
progress to the user.

There is a few follow-up items in the code that I'm not addressing in
this PR – they are related to Blueprints v2-specific code on
playground.wordpress.net so it's fine if they're imperfect.

Supersedes #2663
A part of #2586