[Website] Resolve the Blueprint declaration for the 'View Blueprint' button (#2675)
## Motivation for the change, related issues

Improves the #2562 situation in that it brings over the Blueprint to the
JSON editor when you click "View Blueprint". It's still not a full
solution as the editor doesn't seem to support the "bundled" resources
at the moment, e.g.

```ts
    {
      "step": "writeFile",
      "path": "/wordpress/wp-content/mu-plugins/rewrite.php",
      "data": {
        "resource": "bundled",
        "path": "./scripts/setup.php"
      }
    },
```

Still, it improves the user experience so we might land it and keep
thinking.

cc @bph @brandonpayton 

## Testing Instructions (or ideally a Blueprint)

Follow the steps in #2562 and confirm the Blueprint editor shows the
correct JSON file.