[PHP] Do not reap() PHP in the worker until the streaming response ends (#2703)
## Motivation for the change, related issues

With this PR, PHPWorker only `reap()`s the PHP process once the
`StreamedPHPResponse` is finished. Before this PR, it was `reap()`-ed
immediately after the `StreamedPHPResponse` was created but before all
the data was generated and consumed streamed.

## Testing Instructions (or ideally a Blueprint)

Testing of this PR is not trivial and I gave up on writing a test after
two hours of fiddling around. Let's just confirm it doesn't break any
other test.